Move notebook chat edit contribs into separate dir
