+++ src/vs/workbench/contrib/notebook/browser/contrib/chatEdit/notebookOriginalModelRefFactory.ts
@@ -3,12 +3,12 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { AsyncReferenceCollection, IReference, ReferenceCollection } from '../../../../../base/common/lifecycle.js';
-import { IModifiedFileEntry } from '../../../chat/common/chatEditingService.js';
-import { INotebookService } from '../../common/notebookService.js';
-import { bufferToStream, VSBuffer } from '../../../../../base/common/buffer.js';
-import { NotebookTextModel } from '../../common/model/notebookTextModel.js';
-import { createDecorator, IInstantiationService } from '../../../../../platform/instantiation/common/instantiation.js';
+import { AsyncReferenceCollection, IReference, ReferenceCollection } from '../../../../../../base/common/lifecycle.js';
+import { IModifiedFileEntry } from '../../../../chat/common/chatEditingService.js';
+import { INotebookService } from '../../../common/notebookService.js';
+import { bufferToStream, VSBuffer } from '../../../../../../base/common/buffer.js';
+import { NotebookTextModel } from '../../../common/model/notebookTextModel.js';
+import { createDecorator, IInstantiationService } from '../../../../../../platform/instantiation/common/instantiation.js';
 
 
 export const INotebookOriginalModelReferenceFactory = createDecorator<INotebookOriginalModelReferenceFactory>('INotebookOriginalModelReferenceFactory');
