+++ src/vs/workbench/contrib/notebook/browser/notebook.contribution.ts
@@ -103,6 +103,9 @@ import './contrib/multicursor/notebookMulticursor.js';
 // Diff Editor Contribution
 import './diff/notebookDiffActions.js';
 
+// Chat Edit Contributions
+import './chatEdit/notebookChatEditController.js';
+
 // Services
 import { editorOptionsRegistry } from '../../../../editor/common/config/editorOptions.js';
 import { NotebookExecutionStateService } from './services/notebookExecutionStateServiceImpl.js';
@@ -130,11 +133,6 @@ import { NotebookMultiDiffEditorInput } from './diff/notebookMultiDiffEditorInpu
 import { getFormattedMetadataJSON } from '../common/model/notebookCellTextModel.js';
 import { INotebookOutlineEntryFactory, NotebookOutlineEntryFactory } from './viewModel/notebookOutlineEntryFactory.js';
 import { getFormattedNotebookMetadataJSON } from '../common/model/notebookMetadataTextModel.js';
-import { NotebookChatEditorControllerContrib } from './chatEdit/notebookChatEditController.js';
-import { registerNotebookContribution } from './notebookEditorExtensions.js';
-import { INotebookOriginalModelReferenceFactory, NotebookOriginalModelReferenceFactory } from './chatEdit/notebookOriginalModelRefFactory.js';
-import { INotebookModelSynchronizerFactory, NotebookModelSynchronizerFactory } from './chatEdit/notebookSynchronizer.js';
-import { INotebookOriginalCellModelFactory, OriginalNotebookCellModelFactory } from './chatEdit/notebookOriginalCellModelFactory.js';
 
 /*--------------------------------------------------------------------------------------------- */
 
@@ -881,10 +879,6 @@ registerSingleton(INotebookLoggingService, NotebookLoggingService, Instantiation
 registerSingleton(INotebookCellOutlineDataSourceFactory, NotebookCellOutlineDataSourceFactory, InstantiationType.Delayed);
 registerSingleton(INotebookOutlineEntryFactory, NotebookOutlineEntryFactory, InstantiationType.Delayed);
 
-registerNotebookContribution(NotebookChatEditorControllerContrib.ID, NotebookChatEditorControllerContrib);
-registerSingleton(INotebookOriginalModelReferenceFactory, NotebookOriginalModelReferenceFactory, InstantiationType.Delayed);
-registerSingleton(INotebookModelSynchronizerFactory, NotebookModelSynchronizerFactory, InstantiationType.Delayed);
-registerSingleton(INotebookOriginalCellModelFactory, OriginalNotebookCellModelFactory, InstantiationType.Delayed);
 
 const schemas: IJSONSchemaMap = {};
 function isConfigurationPropertySchema(x: IConfigurationPropertySchema | { [path: string]: IConfigurationPropertySchema }): x is IConfigurationPropertySchema {
